Add functionality for updating polylines upon marker changes
Modify the updateMarkerLocation method to update the coordinates any 
polylines that use the corresponding marker's location. This will allow 
the polyline to update whenever a marker is dragged.
Modify the deleteMarker method to remove the coordinates that a marker 
is associated with when it is deleted (This will remove any coordinates 
from a polyline when their corresponding markers are deleted).
Polylines are currently not saved to the device. There currently is no 
way to delete them.