Add functionality for pressing the polylines.
Click on polyline to select it. This will cause the polyline to turn red 
and its markers to turn green. 
Click on a selected polyline or another polyline, to deselect it. This 
will cause it to revert to its unselected form (black, red markers).
The polyline button will still deselect the current polyline. If no 
polylines are selected, it will create a new polyline.